Test was moved to ATest page
