+++ src/test/java/tests/browse_languages/letters/ATest.java
@@ -1,11 +1,13 @@
 package tests.browse_languages.letters;
 
 import base.BaseTest;
+import org.openqa.selenium.WebElement;
 import org.testng.Assert;
 import org.testng.annotations.Test;
 import pages.browse_languages.languages.AdaLanguagePage;
 import pages.browse_languages.letters.APage;
 
+import java.util.Arrays;
 import java.util.List;
 
 public class ATest extends BaseTest {
@@ -49,5 +51,30 @@ public class ATest extends BaseTest {
         Assert.assertNotEquals(oldURL, newURL);
         Assert.assertEquals(adaLanguagePage.getURL(), newURL);
         Assert.assertEquals(adaLanguagePage.getTitle(), actualTitle);
+    }
+
+
+    @Test
+    public void testABCPagesLanguagesNameAndInformationColumn() {
+
+        final List<String> expectedInfoColumn = Arrays.asList("Date:", "Author:", "URL:", "Comments:", "Info:", "Score:");
+        APage apage = new APage(getDriver());
+        List<WebElement> languagesSubmenus =
+                openBaseURL()
+                        .clickBrowseLanguagesMenu()
+                        .clickASubmenu()
+                        .getLanguagesLinks();
+        List<String> submenusNames = apage.getNames();
+
+        for (int i = 0; i < languagesSubmenus.size(); i++) {
+            languagesSubmenus.get(i).click();
+            String actualH2Header = apage.getH2HeaderText();
+            String expectedH2Header = "Language " + submenusNames.get(i);
+            List<String> actualInfoColumn = apage.languageInfoDates();
+            apage.clickASubmenu();
+
+            Assert.assertEquals(actualInfoColumn, expectedInfoColumn);
+            Assert.assertEquals(actualH2Header, expectedH2Header);
+        }
     }
 }
